bounrary and occurrence search working
