+++ app/src/components/map/components/StaticLayers.tsx
@@ -28,7 +28,7 @@ export interface IStaticLayerMap {
 export interface IStaticLayer {
   layerName: string;
   features: IStaticLayerFeature[];
-  visible: boolean;
+  dataset_id: string;
 }
 
 export interface IStaticLayersProps {
@@ -48,7 +48,7 @@ const StaticLayers: React.FC<React.PropsWithChildren<IStaticLayersProps>> = (pro
     }
 
     layerControls.push(
-      <LayersControl.Overlay checked={layer.visible} name={layer.layerName} key={`static-layer-${(Math.random() + 1).toString(36).substring(7)}`}>
+      <LayersControl.Overlay checked name={layer.layerName} key={`static-layer-${(Math.random() + 1).toString(36).substring(7)}`}>
         <FeatureGroup key={`static-feature-group-${(Math.random() + 1).toString(36).substring(7)}`}>
           {layer.features.map((item, index) => {
             const id = item.key || item.geoJSON.id || index;
