+++ app/src/features/datasets/components/SearchResultProjectList.tsx
@@ -5,31 +5,49 @@ import Checkbox from '@mui/material/Checkbox';
 import React, { useEffect, useState } from 'react';
 import { Container } from '@mui/system';
 import { Button } from '@mui/material';
-import { groupSpatialDataBySpecies, ISpatialDataGroupedBySpecies } from 'utils/spatial-utils';
-import { IDatasetVisibility, ISearchResultListProps } from './SearchResultOccurrenceList';
+import { DataLoader } from 'hooks/useDataLoader';
+import { ISpatialData } from 'interfaces/useSearchApi.interface';
+import { Feature, GeoJsonProperties, Geometry } from 'geojson';
 
+export interface ISearchResultListProps {
+    mapDataLoader: DataLoader<
+    [
+      searchBoundary: Feature<Geometry, GeoJsonProperties>[],
+      searchType: string[],
+      species?: string[],
+      searchZoom?: number,
+      datasetID?: string
+    ],
+    ISpatialData[],
+    unknown
+    >;
+    onToggleDataVisibility: (datasets: IDatasetVisibility) => void;
+    backToSearch: () => void;
+}
+
+export interface IDatasetVisibility {
+    [details: string]: boolean
+}
+
+// this should be two components, one for the occurence and one for projects
 const SearchResultProjectList: React.FC<ISearchResultListProps> = (props) => {
-    const [groupedSpatialData, setGroupedSpatialData] = useState<ISpatialDataGroupedBySpecies>({})
     const [datasetVisibility, setDatasetVisibility] = useState<IDatasetVisibility>({})
     const {mapDataLoader} = props
 
-    
     useEffect(() => {
         if (!mapDataLoader.data) {
             return;
         }
-        console.log(mapDataLoader.data)
-        const groupedData = groupSpatialDataBySpecies(mapDataLoader.data);
         
-        const setup = {}
-        for(const key in groupedData) {
-            setup[key] = true
-        }
+        const setup = {};
+        mapDataLoader.data.forEach(item => {
+            setup[item.submission_spatial_component_id] = true
+        })
         setDatasetVisibility(setup)
-        setGroupedSpatialData(groupedData)
-    }, [props.mapDataLoader.data])
+    }, [mapDataLoader.data])
 
     const toggleVisibility = (dataset_id: string) => {
+        console.log("TOGGLE VISIBLITY FOR: " + dataset_id)
         const udpated = datasetVisibility
         const value = datasetVisibility[dataset_id];
         udpated[dataset_id] = !udpated[dataset_id]
@@ -37,24 +55,12 @@ const SearchResultProjectList: React.FC<ISearchResultListProps> = (props) => {
         props.onToggleDataVisibility(udpated);
     }
 
-    const countGrouped = (groupedData: ISpatialDataGroupedBySpecies) => {
-        let count = 0;
-        for (const key in groupedData) {
-            const item = groupedData[key]
-            if (item) {
-                count += item.length
-            }
-        }
-
-        return count;
-    }
-
     return (
     <>
         <Box mb={3} flexDirection={"column"}>
             <Grid item xs={8}>
                 <Typography variant="h6">
-                    Found {countGrouped(groupedSpatialData)} observations
+                    Found {mapDataLoader.data?.length} observations
                 </Typography>
             </Grid>
             <Grid item xs={4}>
@@ -68,23 +74,24 @@ const SearchResultProjectList: React.FC<ISearchResultListProps> = (props) => {
         <Container maxWidth="xl">
             <Box>
                 <Grid container direction={"column"} justifyContent="center">
-                    {Object.keys(groupedSpatialData).map((key: string, index: number) => {
+                    {mapDataLoader.data?.map((item: ISpatialData, index: number) => {
+                        console.log(`____ FIRST FUN PLEASE: ${datasetVisibility[item.submission_spatial_component_id] === undefined ? true : datasetVisibility[item.submission_spatial_component_id]}`);
                         return (
-                            <Grid container direction="row" alignItems={"center"} key={`${key}-${index}`}>
+                            <Grid container direction="row" alignItems={"center"} key={`${item.submission_spatial_component_id}-${index}`}>
                                 <Grid item xs={3}>
                                     <Checkbox
-                                        checked={datasetVisibility[key] == undefined ? true : datasetVisibility[key]}
-                                        onChange={() => toggleVisibility(key)}
+                                        checked={datasetVisibility[item.submission_spatial_component_id] === undefined ? true : datasetVisibility[item.submission_spatial_component_id]}
+                                        onChange={() => toggleVisibility(`${item.submission_spatial_component_id}`)}
                                     />
                                 </Grid>
                                 <Grid item xs={4}>
                                     <Typography variant="body1" color="textPrimary">
-                                        {key}
+                                        {item.spatial_data.features[0]?.properties?.datasetTitle}
                                     </Typography>
                                 </Grid>
                                 <Grid item xs={5}>
                                     <Typography variant="body1" color="textPrimary">
-                                        {groupedSpatialData[key].length} records
+                                        {0} records
                                     </Typography>
                                 </Grid>
                             </Grid>
